+++ application/models/Tasks.php
@@ -80,6 +80,15 @@ class Tasks extends XML_Model {
 		$this->rest->option(CURLOPT_PORT, REST_PORT);
 		return $this->rest->get('/job/' . $key);
 	}
+
+	// Delete a record from the DB
+	function delete($key, $key2 = null)
+	{
+		$this->rest->initialize(array('server' => REST_SERVER));
+		$this->rest->option(CURLOPT_PORT, REST_PORT);
+		$this->rest->delete('/job/' . $key);
+		$this->load(); // because the "database" might have changed
+	}
 }
 
 // return -1, 0, or 1 of $a's category name is earlier, equal to, or later than $b's